factor out API into a separate module
